Refactor validation & prep for PR
